Fix: Move npm-artifafctory-shim into gradle-typescript project
