+++ gradle-typescript/build.gradle
@@ -25,7 +25,6 @@ dependencies {
     compile localGroovy()
     compile gradleApi()
 
-    implementation project(':npm-artifactory-shim')
     implementation 'com.fasterxml.jackson.core:jackson-databind'
     implementation 'com.fasterxml.jackson.datatype:jackson-datatype-guava'
     implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jdk8'
