+++ gradle-typescript/src/main/java/com/gradlets/gradle/typescript/shim/ModuleIdentifier.java
