+++ gradle-typescript/src/main/java/com/gradlets/gradle/typescript/shim/cache/CachedDescriptor.java
@@ -16,7 +16,7 @@
 
 package com.gradlets.gradle.typescript.shim.cache;
 
-import com.gradlets.gradle.typescript.shim.ImmutablesStyle;
+import com.gradlets.gradle.ImmutablesStyle;
 import com.gradlets.gradle.typescript.shim.clients.PackageJson;
 import java.util.function.Supplier;
 import org.immutables.value.Value;
