+++ gradle-typescript/src/test/java/com/gradlets/gradle/typescript/shim/EnvironmentVariablesExtension.java
