--- npm-artifactory-shim/build.gradle
@@ -1,37 +1,0 @@
-/*
- * (c) Copyright 2021 Felipe Orozco, Robert Kruszewski. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-apply plugin: 'application'
-
-application {
-    mainClassName = 'com.gradlets.gradle.typescript.shim.NpmArtifactoryShim'
-}
-
-dependencies {
-    implementation 'com.fasterxml.jackson.core:jackson-databind'
-    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-guava'
-    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jdk8'
-    implementation 'com.palantir.dialogue:dialogue-clients'
-    implementation 'com.palantir.dialogue:dialogue-target'
-    implementation 'com.palantir.safe-logging:preconditions'
-    implementation 'io.undertow:undertow-core'
-
-    testImplementation 'org.assertj:assertj-core'
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-
-    annotationProcessor 'org.immutables:value'
-    compileOnly 'org.immutables:value::annotations'
-}
