+++ settings.gradle
@@ -19,4 +19,3 @@ rootProject.name = 'gradle-typescript-root'
 include 'gradle-baseline-typescript'
 include 'gradle-conjure-typescript'
 include 'gradle-typescript'
-include 'npm-artifactory-shim'
